requirements.txt
# requirements.txt — Python packages needed for this project
#
# Install all of these by running:
#   pip install -r requirements.txt

# For making HTTP requests (fetching job data from APIs)
requests==2.31.0

# For parsing HTML when scraping web pages
beautifulsoup4==4.12.3

# For parsing HTML (used by BeautifulSoup)
lxml==5.1.0

# For loading environment variables from a .env file (local development only)
python-dotenv==1.0.1
